+++ web/src/App.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Footer from "./components/Footer";
+import Footer from "./components/navigation/Footer";
 
 const App = () => {
   return (
