+++ web/src/components/events-slider/UpcomingEventCard.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Event } from "../types/types";
+import { Event } from "../../types/types";
 import { FaCalendarAlt } from "react-icons/fa";
 import { FaLocationDot } from "react-icons/fa6";
 import dayjs from "dayjs";
