+++ web/src/components/exec-page/ExecCard.tsx
@@ -1,4 +1,4 @@
-import { ExecCardProps } from "../types/types";
+import { ExecCardProps } from "../../types/types";
 
 export default function ExecCard({ exec }: ExecCardProps) {
   return (
