+++ web/src/components/exec-page/PreviousTeamCard.tsx
@@ -1,4 +1,4 @@
-import type { PreviousTeamCardProps } from "../types/types";
+import type { PreviousTeamCardProps } from "../../types/types";
 
 export default function PreviousTeamCard({
   year,
